+++ cli/src/vscodeserver.ts
@@ -1,0 +1,6 @@
+export async function vscodeserver(
+    fn: string | undefined,
+    options: object = {}
+) {
+    
+}
